requirements.txt
aiogram~=3.10
psycopg2-binary
python-dotenv
httpx